limpieza de funciones de extraccion
ya @alexguz
