+++ src/Components/Modal.js
@@ -21,7 +21,7 @@ const Modal = ({isOpen, close }) => {
 
     try {
       setIsLoading(true);
-      const uploadResponse = await fetch('http://localhost:5000/U1/upload_file2', {
+      const uploadResponse = await fetch('http://localhost:5000/U1/analyzeDoc', {
       //const uploadResponse = await fetch('https://fridarender.onrender.com/upload_file', {
         method: 'POST',
         body: formData,
